feat: Add ABTestModelFactory and related classes
This commit adds the ABTestModelFactory class along with the BetaPrior and BinomialData classes. These classes are used for creating Bayesian A/B test models. The ABTestModelFactory class has a method create_model() that takes a list of BinomialData objects and returns a PyMC3 model for the A/B test. This commit also includes the necessary imports and dependencies.